Изменения в админке - категории, категория, проект, проекты - обновление install.mysql.utf8.sql поле gallery - во вьюхи добавлено разное отображение кнопок на тулбаре в зависимости от версии Joomla
