+++ com_swjprojects/admin/views/projects/tmpl/default.php
@@ -10,159 +10,6 @@
 
 defined('_JEXEC') or die;
 
-use Joomla\CMS\Factory;
-use Joomla\CMS\HTML\HTMLHelper;
-use Joomla\CMS\Language\Text;
-use Joomla\CMS\Layout\LayoutHelper;
-use Joomla\CMS\Router\Route;
+use Joomla\CMS\Version;
 
-HTMLHelper::_('jquery.framework');
-HTMLHelper::_('behavior.formvalidator');
-HTMLHelper::_('behavior.keepalive');
-HTMLHelper::_('behavior.tabstate');
-HTMLHelper::_('formbehavior.chosen', 'select');
-HTMLHelper::stylesheet('com_swjprojects/admin-j3.min.css', array('version' => 'auto', 'relative' => true));
-
-$user      = Factory::getUser();
-$listOrder = $this->escape($this->state->get('list.ordering'));
-$listDirn  = $this->escape($this->state->get('list.direction'));
-$saveOrder = ($listOrder == 'p.ordering' && strtolower($listDirn) == 'asc');
-
-if ($saveOrder)
-{
-	$saveOrderingUrl = 'index.php?option=com_swjprojects&task=projects.saveOrderAjax&tmpl=component';
-	HTMLHelper::_('sortablelist.sortable', 'projectsList', 'adminForm', strtolower($listDirn), $saveOrderingUrl);
-}
-
-$columns = 8;
-?>
-<form action="<?php echo Route::_('index.php?option=com_swjprojects&view=projects'); ?>" method="post"
-	  name="adminForm" id="adminForm">
-	<div id="j-sidebar-container" class="span2">
-		<?php echo $this->sidebar; ?>
-	</div>
-	<div id="j-main-container" class="span10">
-		<?php echo LayoutHelper::render('joomla.searchtools.default', array('view' => $this)); ?>
-		<?php if (empty($this->items)) : ?>
-			<div class="alert alert-no-items">
-				<?php echo Text::_('JGLOBAL_NO_MATCHING_RESULTS'); ?>
-			</div>
-		<?php else : ?>
-			<table id="projectsList" class="table table-striped">
-				<thead>
-				<tr>
-					<th width="1%" class="nowrap center hidden-phone">
-						<?php echo HTMLHelper::_('searchtools.sort', '', 'p.ordering', $listDirn, $listOrder, null,
-							'asc', 'JGRID_HEADING_ORDERING', 'icon-menu-2'); ?>
-					</th>
-					<th width="1%" class="center">
-						<?php echo HTMLHelper::_('grid.checkall'); ?>
-					</th>
-					<th width="2%" style="min-width:100px" class="center">
-						<?php echo HTMLHelper::_('searchtools.sort', 'JSTATUS', 'p.state', $listDirn, $listOrder); ?>
-					</th>
-					<th class="nowrap">
-						<?php echo HTMLHelper::_('searchtools.sort', 'JGLOBAL_TITLE', 'title', $listDirn, $listOrder); ?>
-					</th>
-					<th width="10%" class="nowrap hidden-phone">
-						<?php echo HTMLHelper::_('searchtools.sort', 'COM_SWJPROJECTS_CATEGORY', 'category_title',
-							$listDirn, $listOrder); ?>
-					</th>
-					<th width="10%" class="nowrap hidden-phone center">
-						<?php echo HTMLHelper::_('searchtools.sort', 'COM_SWJPROJECTS_DOWNLOAD_TYPE', 'p.download_type',
-							$listDirn, $listOrder); ?>
-					</th>
-					<th width="10%" class="nowrap hidden-phone center">
-						<?php echo HTMLHelper::_('searchtools.sort', 'COM_SWJPROJECTS_STATISTICS_DOWNLOADS', 'downloads',
-							$listDirn, $listOrder); ?>
-					</th>
-					<th width="10%" class="nowrap hidden-phone center">
-						<?php echo HTMLHelper::_('searchtools.sort', 'COM_SWJPROJECTS_STATISTICS_HITS', 'p.hits',
-							$listDirn, $listOrder); ?>
-					</th>
-					<th width="1%" class="nowrap hidden-phone center">
-						<?php echo HTMLHelper::_('searchtools.sort', 'JGRID_HEADING_ID', 'p.id', $listDirn, $listOrder); ?>
-					</th>
-				</tr>
-				</thead>
-				<tfoot>
-				<tr>
-					<td colspan="<?php echo $columns; ?>">
-						<?php echo $this->pagination->getListFooter(); ?>
-					</td>
-				</tr>
-				</tfoot>
-				<tbody>
-				<?php foreach ($this->items as $i => $item) :
-					$canEdit = $user->authorise('core.edit', 'com_swjprojects.project.' . $item->id);
-					$canChange = $user->authorise('core.edit.state', 'com_swjprojects.project.' . $item->id);
-					?>
-					<tr class="row<?php echo $i % 2; ?>" item-id="<?php echo $item->id ?>">
-						<td class="order nowrap center hidden-phone">
-							<?php
-							$iconClass = '';
-							if (!$canChange)
-							{
-								$iconClass = ' inactive';
-							}
-							elseif (!$saveOrder)
-							{
-								$iconClass = ' inactive tip-top hasTooltip" title="' .
-									HTMLHelper::_('tooltipText', 'JORDERINGDISABLED');
-							}
-							?>
-							<span class="sortable-handler<?php echo $iconClass ?>"><span
-										class="icon-menu"></span></span>
-							<?php if ($canChange && $saveOrder) : ?>
-								<input type="text" name="order[]" value="<?php echo $item->ordering; ?>"
-									   class="width-20 text-area-order" style="display:none"/>
-							<?php endif; ?>
-						</td>
-						<td class="center">
-							<?php echo HTMLHelper::_('grid.id', $i, $item->id); ?>
-						</td>
-						<td class="center nowrap">
-							<div class="btn-group">
-								<?php echo HTMLHelper::_('jgrid.published', $item->state, $i, 'projects.', $canChange); ?>
-							</div>
-						</td>
-						<td class="nowrap">
-							<?php if ($canEdit) : ?>
-								<a class="hasTooltip" title="<?php echo Text::_('JACTION_EDIT'); ?>"
-								   href="<?php echo Route::_('index.php?option=com_swjprojects&task=project.edit&id='
-									   . $item->id); ?>">
-									<?php echo $this->escape($item->title); ?>
-								</a>
-							<?php else : ?>
-								<?php echo $this->escape($item->title); ?>
-							<?php endif; ?>
-						</td>
-						<td class="small hidden-phone">
-							<?php echo Text::_($this->escape($item->category_title)); ?>
-						</td>
-						<td class="hidden-phone center">
-							<?php echo Text::_('COM_SWJPROJECTS_DOWNLOAD_TYPE_' . $item->download_type); ?>
-						</td>
-						<td class="hidden-phone center">
-							<span class="badge badge-info">
-								<?php echo (int) $item->downloads; ?>
-							</span>
-						</td>
-						<td class="hidden-phone center">
-							<span class="badge badge-info">
-								<?php echo (int) $item->hits; ?>
-							</span>
-						</td>
-						<td class="hidden-phone center">
-							<?php echo $item->id; ?>
-						</td>
-					</tr>
-				<?php endforeach; ?>
-				</tbody>
-			</table>
-		<?php endif; ?>
-		<input type="hidden" name="task" value=""/>
-		<input type="hidden" name="boxchecked" value="0"/>
-		<?php echo HTMLHelper::_('form.token'); ?>
-	</div>
-</form>
+echo $this->loadTemplate(((new Version())->isCompatible('4.0')) ? 'j4' : 'j3');
