+++ com_swjprojects/media/css/admin-j4.css
@@ -29,4 +29,6 @@
 	font-weight: 900 !important;
 }
 
-
+.subhead #toolbar.btn-toolbar a[target=_blank]:before {
+	content: '';
+}
